==> code/analysis/1-action-consort.Rmd <==
---
title: "ACTION trial"
subtitle: "CONSORT Participant Flow"
author: 
- Ben Arnold ben.arnold@ucsf.edu 
- Fanice Nyatigo fanice.nyatigo@ucsf.edu
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: default
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
knit: (function(inputFile, encoding) {
        rmarkdown::render(inputFile, 
                          encoding   = encoding, 
                          output_dir = "../../output/analysis"
                          )})
---

# Summary

Track participant flow in the ACTION trial.

Summarize patient numbers by group to include in a CONSORT flow diagram.

# Preamble / configuration

```{r preamble, message = FALSE}
#---------------------------------
# source the project's configuration
# file
#---------------------------------
library(here)
here()
source(here("code","config.R"))


```


# Load the data

Load the final population tracking dataset. This file was created by`ACTION_analysis_data_prep.R`.


```{r load the data}
#---------------------------------
# load the preprocessed data

# final public study population tracking dataset
# created with ACTION_analysis_data_prep.R
# and ACTION_create_public_datasets.R
#---------------------------------
dc <- read_rds(here("data/public","ACTION_consort_public.rds"))

#---------------------------------
# rename the public id to recordID
#---------------------------------

dc %<>% rename(recordID = id_public)

#---------------------------------
# create a shorter named treatment
# variable 
#---------------------------------
dc %<>%
  mutate(tr = tr_received)
```

# CONSORT flow

Participant flow by study visit

Lost to follow-up is a mutually exclusive outcome across study visits, whereas treatment and contributing to the analysis is not. That makes the table include redundant "No" levels for each of the treated and analyzed entries which we can ignore when transferring this information into the CONSORT diagram.

## Screening and Enrollment
```{r consort flow screening}
#---------------------------------
# Screening and Enrollment
#---------------------------------
dc1 <- dc %>% summarise(
  `Assessed for eligibility`                 = n(),
  `Excluded (Overall)`                       = sum(consented %in% c(2,3) | is.na(consented)),
  `COVID test out of window`                 = sum(elig_window == 0, na.rm = T),
  `Negative COVID test`                      = sum(elig_test == 0, na.rm = T),
  `Could not be reached for consent`         = sum(consented == 3, na.rm = T),
  `Refused (Overall)`                        = sum(consented == 2, na.rm = T),
  `DOB implies below 18 years old`           = sum(elig_age == 0, na.rm = T),
  `Unable to complete online questionnaires` = sum(elig_onlineforms == 0, na.rm = T),
  `Unable to receive study drug in mail`     = sum(elig_maildrug == 0, na.rm = T),
  `Macrolide allergy`                        = sum(elig_allergy == 0, na.rm = T),
  `Over 55 and on hydroxychloroquine`        = sum(elig_hcq55 == 0, na.rm = T),
  `Currently taking nelfinavir or warfarin`  = sum(elig_nelfwar == 0, na.rm = T),
  `Prolonged QT`                             = sum(elig_qt == 0, na.rm = T),
  `Pregnant`                                 = sum(elig_preg == 0, na.rm = T),
  `Other reason for ineligibility`           = sum(elig_no_other == 0, na.rm = T),
  `Enrolled (Overall)`                       = sum(consented == 1, na.rm = T)
) %>% 
  # transpose
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "Variable") %>% 
  rename(Number = V1) 

kable(dc1) %>% 
  row_spec(c(1, 2, 16), bold = T, italic = T)
```
*Note: in response to editorial comments we included a finer stratification of the refusals and other reasons for ineligibility, tallied by hand, for the final flow diagram*

## Allocation and Follow-up
```{r consort flow followup}
#---------------------------------
# among participants enrolled
# identify treatment and 
# participants lost to follow-up 
# at various points
#---------------------------------
dc2 <- dc %>%
  # restrict to participants enrolled
  filter(consented == 1) %>%
  # create label for mismatch in tr
  #mutate() %>% 
  # create lost to follow-up categorical var
  mutate(lost_to_fu = case_when(
    ltfu3 == 1 ~ "Lost days 0-3",
    ltfu7 == 1 ~ "Lost days 3-7",
    ltfu14 == 1 ~ "Lost days 7-14",
    ltfu21 == 1 ~ "Lost days 14-21",
    TRUE ~ as.character("Completed follow-up through day 21")
    ),
    lost_to_fu = factor(lost_to_fu, levels = c("Lost days 0-3","Lost days 3-7","Lost days 7-14","Lost days 14-21","Completed follow-up through day 21"))
    ) %>%
  # recode analysis population variables as Y/N
  mutate(
    an3 = factor(an3, levels=c(0,1), labels = c("No","Yes")),
    an7 = factor(an7, levels=c(0,1), labels = c("No","Yes")),
    an14 = factor(an14, levels=c(0,1), labels = c("No","Yes")),
    an21 = factor(an21, levels=c(0,1), labels = c("No","Yes")),
  )

#---------------------------------
# placeholder for treatment status
#---------------------------------


#---------------------------------
# label the variables for table
# printing
#---------------------------------

label(dc2$take_tr) <- "Treatment Administered"
label(dc2$lost_to_fu) <- "Lost to follow-up"
label(dc2$an3) <- "Analyzed, day 3"
label(dc2$an7) <- "Analyzed, day 7"
label(dc2$an14) <- "Analyzed, day 14"
label(dc2$an21) <- "Analyzed, day 21"

table1(~ take_tr + lost_to_fu + an3 + an7 + an14 + an21 | tr, data = dc2)
```

# Session Info
```{r session info}
sessionInfo()
```
